Update add image wizard panels
