+++ Core/src/org/sleuthkit/autopsy/casemodule/AddImageVisualPanel3.java
@@ -39,8 +39,6 @@ public class AddImageVisualPanel3 extends javax.swing.JPanel {
         configPanel.setBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(0, 0, 0)));
         configPanel.setLayout(new BorderLayout());
         configPanel.add(ingestPanel, BorderLayout.CENTER);
-        
-        
     }
     
      /**
@@ -97,7 +95,7 @@ public class AddImageVisualPanel3 extends javax.swing.JPanel {
                     .addGroup(layout.createSequentialGroup()
                         .addContainerGap()
                         .addComponent(subtitleLabel)))
-                .addContainerGap(132, Short.MAX_VALUE))
+                .addContainerGap(247, Short.MAX_VALUE))
             .addComponent(configPanel, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
         );
         layout.setVerticalGroup(
