+++ Core/src/org/sleuthkit/autopsy/casemodule/Bundle.properties
@@ -87,9 +87,6 @@ CueBannerPanel.openCaseButton.text=
 CueBannerPanel.openRecentButton.text=
 OpenRecentCasePanel.cancelButton.text=Cancel
 OpenRecentCasePanel.jLabel1.text=Recent Cases
-AddImageVisualPanel2.crDbLabel.text=Adding Image
-AddImageVisualPanel2.jLabel5.text=Processing Image and Adding to Database
-AddImageVisualPanel2.jLabel1.text=<html> We are now analyzing the disk image to extract volume and file system data and populate a local database. </html>
 NewJPanel.jLabel1.text=
 NewJPanel.jLabel2.text=NSRL Index
 NewJPanel.jLabel5.text=The NSRL index is used to idenify "known" files.
@@ -101,7 +98,7 @@ AddImageVisualPanel4.jLabel1.text=Image was successfully added to the case and i
 AddImageVisualPanel4.crDbLabel.text=Image Added and Being Processed 
 AddImageVisualPanel4.addImgButton.text=Add Another Image
 AddImageVisualPanel3.titleLabel.text=Configure Ingest Modules
-AddImageVisualPanel3.subtitleLabel.text=Image has been commited.  You can configure and run ingest modules on the new image.
+AddImageVisualPanel3.subtitleLabel.text=Configure the ingest modules you would like to run on this image.
 CasePropertiesForm.caseNumberLabel.text=Case Number:
 CasePropertiesForm.examinerLabel.text=Examiner:
 CasePropertiesForm.caseNumberTextField.text=
@@ -135,4 +132,9 @@ MissingImageDialog.cancelButton.text=Cancel
 LocalDiskPanel.errorLabel.text=Error Label
 AddImageVisualPanel1.typeTabel.text=Select input type to add:
 AddImageVisualPanel1.imgInfoLabel.text=Enter Disk Image Information:
-AddImageVisualPanel2.Label_CurrentDirectory_Static.text=Currently Processing:
+AddImageLoadingPanel.crDbLabel.text=Adding Image
+AddImageLoadingPanel.Label_CurrentDirectory_Static.text=Currently Processing:
+AddImageLoadingPanel.jLabel5.text=Processing Image and Adding to Database
+AddImageLoadingPanel.jLabel1.text=File system information is being added to a local database. File analysis will start when this finishes.
+AddImageDonePanel.statusLabel.text=File system has been added to the local database. Files are being analyzed.
+AddImageDonePanel.crDbLabel.text=Adding Image - Complete
